+++ src/components/CodeDiff/index.js
@@ -1,9 +1,2 @@
 import CodeDiff from './CodeDiff';
-import CodeDiffLegacy from './CodeDiffLegacy';
-
-// TODO(APPX-1337): Remove after new CodeDiff is determined to be stable
-export {
-  CodeDiff,
-  CodeDiffLegacy,
-};
 export default CodeDiff;